requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
distlib==0.3.8
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.13.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==5.1.0
mysql-connector-python==8.3.0
openai==1.14.1
packaging==24.0
parsel==1.8.1
pillow==10.2.0
platformdirs==4.2.0
Protego==0.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyDispatcher==2.0.7
PyMySQL==1.1.0
pyOpenSSL==24.1.0
queuelib==1.6.2
requests==2.31.0
requests-file==2.0.0
Scrapy==2.11.1
scrapy-user-agents==0.1.1
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
tldextract==5.1.1
tqdm==4.66.2
Twisted==24.3.0
typing_extensions==4.10.0
ua-parser==0.18.0
urllib3==2.2.1
user-agents==2.2.0
virtualenv==20.25.1
w3lib==2.1.2
zope.interface==6.2
